server/pyproject.toml
[project]
name = "sc-flask"
version = "0.1.0"
description = "The backend python implementation for SuttaCentral.net"
readme = "README.md"
requires-python = "==3.11.3"
dependencies = [
    "aksharamukha==2.3",
    "alabaster==0.7.12",
    "algoliasearch==3.0.0",
    "aniso8601==9.0.1",
    "appdirs==1.4.4",
    "attrs==22.1.0",
    "babel==2.11.0",
    "beautifulsoup4==4.11.1",
    "certifi==2022.12.7",
    "charset-normalizer==2.1.1",
    "click==8.1.3",
    "cssselect==1.2.0",
    "decorator==5.1.1",
    "deprecated==1.2.13",
    "docutils==0.19",
    "faker==15.3.4",
    "flasgger==0.9.5",
    "flask==2.2.5",
    "flask-caching==2.0.1",
    "flask-cors==3.0.10",
    "flask-restful==0.3.9",
    "fonttools==4.38.0",
    "gitdb==4.0.10",
    "gitpython==3.1.30",
    "idna==3.4",
    "imagesize==1.4.1",
    "importlib-metadata==5.1.0",
    "importlib-resources==5.10.0",
    "iniconfig==1.1.1",
    "itsdangerous==2.1.2",
    "jaconv==0.3",
    "jinja2==3.1.2",
    "jsonschema==4.17.3",
    "langcodes==3.3.0",
    "langid==1.1.6",
    "language-data==1.1",
    "lxml==4.9.1",
    "marisa-trie==0.7.8",
    "markupsafe==2.1.1",
    "mistune==2.0.4",
    "packaging==21.3",
    "pluggy==1.0.0",
    "py==1.11.0",
    "pyee==8.2.2",
    "pyenchant==3.2.2",
    "pygments==2.13.0",
    "pyjwt==2.6.0",
    "pykakasi==2.2.1",
    "pyparsing==3.0.9",
    "pyppeteer==1.0.2",
    "pyrsistent==0.19.2",
    "pytest==7.2.0",
    "pytest-flask==1.2.0",
    "pytest-sugar==0.9.6",
    "python-arango==7.5.3",
    "python-dateutil==2.8.2",
    "pytz==2022.6",
    "pyyaml==6.0",
    "regex==2022.10.31",
    "requests==2.31.0",
    "requests-toolbelt==0.10.1",
    "setuptools-scm==7.0.5",
    "six==1.16.0",
    "smmap==5.0.0",
    "snowballstemmer==2.2.0",
    "sortedcontainers==2.4.0",
    "soupsieve==2.3.1",
    "sphinx==5.3.0",
    "sphinxcontrib-applehelp==1.0.2",
    "sphinxcontrib-devhelp==1.0.2",
    "sphinxcontrib-htmlhelp==2.0.0",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==1.0.3",
    "sphinxcontrib-serializinghtml==1.1.5",
    "sphinxcontrib-websupport==1.2.4",
    "stripe==6.4.0",
    "termcolor==2.1.1",
    "text-unidecode==1.3",
    "toml==0.10.2",
    "tomli==2.0.1",
    "tqdm==4.64.1",
    "typing-extensions==4.4.0",
    "unidecode==1.3.7",
    "urllib3==1.26.17",
    "uwsgi==2.0.21",
    "websockets==10.4",
    "werkzeug==2.3.4",
    "wrapt==1.14.1",
    "zhconv==1.4.3",
    "zipp==3.11.0",
]

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"
